refactor(juicy section): move btn in separate component
